Modified hydrate me module
