+++ .env
@@ -1,2 +1,2 @@
- REACT_APP_API_BASE_URL=http://34.93.134.127/api
-# REACT_APP_API_BASE_URL=http://localhost:8080/api
+# REACT_APP_API_BASE_URL=http://34.93.134.127/api
+REACT_APP_API_BASE_URL=http://localhost:8080/api
